loging sort of working
